refactor: updates config structure for modularity
