+++ src/config/show_details.rs
@@ -1,0 +1,127 @@
+use clap::Parser;
+use serde::Deserialize;
+use std::{fmt::Display, str::FromStr};
+
+#[derive(
+    Deserialize, Debug, Default, clap::Parser, Copy, Clone, clap::ValueEnum, PartialEq, Eq,
+)]
+pub enum ShowCalls {
+    #[default]
+    None,
+    User,
+    System,
+    All,
+}
+
+impl FromStr for ShowCalls {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_ref() {
+            "none" => Ok(ShowCalls::None),
+            "user" => Ok(ShowCalls::User),
+            "system" => Ok(ShowCalls::System),
+            "all" => Ok(ShowCalls::All),
+            _ => Err(format!(
+                "Unknown ShowCalls value {} - expected one of none|user|system|all.",
+                s
+            )),
+        }
+    }
+}
+
+impl Display for ShowCalls {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        write!(f, "{:?}", self)
+    }
+}
+
+#[derive(Deserialize, Debug, Default, Parser, Copy, Clone, clap::ValueEnum, PartialEq, Eq)]
+pub enum ShowStorageLogs {
+    #[default]
+    None,
+    Read,
+    Write,
+    Paid,
+    All,
+}
+
+impl FromStr for ShowStorageLogs {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_ref() {
+            "none" => Ok(ShowStorageLogs::None),
+            "read" => Ok(ShowStorageLogs::Read),
+            "write" => Ok(ShowStorageLogs::Write),
+            "paid" => Ok(ShowStorageLogs::Paid),
+            "all" => Ok(ShowStorageLogs::All),
+            _ => Err(format!(
+                "Unknown ShowStorageLogs value {} - expected one of none|read|write|paid|all.",
+                s
+            )),
+        }
+    }
+}
+
+impl Display for ShowStorageLogs {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        write!(f, "{:?}", self)
+    }
+}
+
+#[derive(Deserialize, Debug, Default, Parser, Copy, Clone, clap::ValueEnum, PartialEq, Eq)]
+pub enum ShowVMDetails {
+    #[default]
+    None,
+    All,
+}
+
+impl FromStr for ShowVMDetails {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_ref() {
+            "none" => Ok(ShowVMDetails::None),
+            "all" => Ok(ShowVMDetails::All),
+            _ => Err(format!(
+                "Unknown ShowVMDetails value {} - expected one of none|all.",
+                s
+            )),
+        }
+    }
+}
+
+impl Display for ShowVMDetails {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        write!(f, "{:?}", self)
+    }
+}
+
+#[derive(Deserialize, Debug, Default, Parser, Copy, Clone, clap::ValueEnum, PartialEq, Eq)]
+pub enum ShowGasDetails {
+    #[default]
+    None,
+    All,
+}
+
+impl FromStr for ShowGasDetails {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_ref() {
+            "none" => Ok(ShowGasDetails::None),
+            "all" => Ok(ShowGasDetails::All),
+            _ => Err(format!(
+                "Unknown ShowGasDetails value {} - expected one of none|all.",
+                s
+            )),
+        }
+    }
+}
+
+impl Display for ShowGasDetails {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        write!(f, "{:?}", self)
+    }
+}
