+++ src/http_fork_source.rs
@@ -5,7 +5,7 @@ use std::{
 
 use crate::{
     cache::Cache,
-    config::CacheConfig,
+    config::cache::CacheConfig,
     fork::{block_on, ForkSource},
 };
 use eyre::Context;
