+++ src/lib.rs
@@ -44,7 +44,6 @@
 pub mod bootloader_debug;
 pub mod config;
 pub mod console_log;
-pub mod constants;
 pub mod deps;
 pub mod filters;
 pub mod fork;
