+++ src/node/fee_model.rs
@@ -1,6 +1,6 @@
 use zksync_types::fee_model::{BatchFeeInput, FeeModelConfigV2, FeeParams, FeeParamsV2};
 
-use crate::config::{
+use crate::config::constants::{
     DEFAULT_ESTIMATE_GAS_PRICE_SCALE_FACTOR, DEFAULT_ESTIMATE_GAS_SCALE_FACTOR,
     DEFAULT_FAIR_PUBDATA_PRICE, DEFAULT_L1_GAS_PRICE, DEFAULT_L2_GAS_PRICE,
 };
@@ -53,35 +53,6 @@ impl TestNodeFeeInputProvider {
             ..Default::default()
         }
     }
-    // Todo: all configuration should be done in config are not here!!
-    // pub fn with_overrides(mut self, gas_config: Option<GasConfig>) -> Self {
-    //     let Some(gas_config) = gas_config else {
-    //         return self;
-    //     };
-
-    //     if let Some(l1_gas_price) = gas_config.l1_gas_price {
-    //         self.l1_gas_price = l1_gas_price;
-    //     }
-
-    //     if let Some(l1_pubdata_price) = gas_config.l1_pubdata_price {
-    //         self.l1_pubdata_price = l1_pubdata_price;
-    //     }
-
-    //     if let Some(l2_gas_price) = gas_config.l2_gas_price {
-    //         self.l2_gas_price = l2_gas_price;
-    //     }
-
-    //     if let Some(estimation) = gas_config.estimation {
-    //         if let Some(factor) = estimation.price_scale_factor {
-    //             self.estimate_gas_price_scale_factor = factor;
-    //         }
-    //         if let Some(factor) = estimation.limit_scale_factor {
-    //             self.estimate_gas_scale_factor = factor;
-    //         }
-    //     }
-
-    //     self
-    // }
 
     pub fn get_fee_model_config(&self) -> FeeModelConfigV2 {
         FeeModelConfigV2 {
